Separate CLI process from commands
