+++ include/pm/ui/cli.h
@@ -2,6 +2,7 @@
 #define PM_UI_CLI_H
 
 #include <pm/ui/cli/commands.h>
+#include <pm/ui/cli/proc.h>
 
 void cli_start(AppState *state);
 void cli_run(AppState *state);
